+++ app/lib/exercise/index.ts
@@ -1,1 +1,2 @@
 export * from './findSimilar';
+export * from './equipment';